Use WebExtension API for FF
