+++ skeletons/web-extension/background-script.js
