+++ skeletons/web-extension/content-script.js
