+++ skeletons/web-extension/devtools.js
