+++ skeletons/web-extension/in-page-script.js
