+++ skeletons/web-extension/options-dialog.html
